requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose
bcrypt
PyJWT
pydantic[email]